all utility functionalities are put into utils.py
